styling

+++ server/public/scripts/client.js
@@ -1,97 +1,93 @@
-console.log( 'js' );
+console.log("js");
 
-$( document ).ready( function(){
-  console.log( 'JQ' );
+$(document).ready(function () {
+  console.log("JQ");
   // Establish Click Listeners
-  setupClickListeners()
+  setupClickListeners();
   // load existing koalas on page load
   getKoalas();
-
 }); // end doc ready
 
 function setupClickListeners() {
-  
-  $( document ).on( 'click', '.deleteBtn', deleteKoala );
-  $( document ).on( 'click', '.transferToggleBtn', toggleReadiness );
+  $(document).on("click", ".deleteBtn", deleteKoala);
+  $(document).on("click", ".transferToggleBtn", toggleReadiness);
 
-
-  $( '#addButton' ).on( 'click', function(){
-    console.log( 'in addButton on click' );
+  $("#addButton").on("click", function () {
+    console.log("in addButton on click");
     // get user input and put in an object
     // NOT WORKING YET :(
     // using a test object
     let koalaToSend = {
-      name: $( '#nameIn' ).val(),
-      age: $( '#ageIn' ).val(),
-      gender: $( '#genderIn' ).val(),
-      readyForTransfer: $( '#readyForTransferIn' ).val(),
-      notes: $( '#notesIn' ).val(),
+      name: $("#nameIn").val(),
+      age: $("#ageIn").val(),
+      gender: $("#genderIn").val(),
+      readyForTransfer: $("#readyForTransferIn").val(),
+      notes: $("#notesIn").val(),
     };
     console.log(koalaToSend);
     // call saveKoala with the new obejct
-    saveKoala( koalaToSend );
-  }); 
+    saveKoala(koalaToSend);
+  });
 }
 
-function getKoalas(){
-  console.log( 'in getKoalas' );
+function getKoalas() {
+  console.log("in getKoalas");
   // ajax call to server to get koalas
   $.ajax({
-    method: 'GET',
-    url: '/koalas'
-  }).then(function(response){
+    method: "GET",
+    url: "/koalas",
+  }).then(function (response) {
     console.log(response);
     showKoalas(response);
-  })
+  });
 } // end getKoalas
 
-function saveKoala( newKoala ){
-  console.log( 'in saveKoala', newKoala );
+function saveKoala(newKoala) {
+  console.log("in saveKoala", newKoala);
   // ajax call to server to post koalas
- $.ajax({
-  
-  method: 'POST',
-  url: '/koalas',
-  data: newKoala
-  
- }).then(function(response) {
-   console.log(response);
+  $.ajax({
+    method: "POST",
+    url: "/koalas",
+    data: newKoala,
+  })
+    .then(function (response) {
+      console.log(response);
+    })
+    .catch(function (error) {
+      console.log("error in koala POST", error);
+      alert("Error adding koala.");
+    });
+} // end saveKoala
 
+function toggleReadiness() {
+  const myID = $(this).data("id");
+  console.log("ID clicked:", myID);
 
- }).catch(function(error) {
-   console.log('error in koala POST', error);
-   alert('Error adding koala.')
- })
-}// end saveKoala
-
-function toggleReadiness (){
-  const myID = $(this).data('id');
-  console.log('ID clicked:', myID);
-  
   let readiness = {
-
-    status: $(this).data('status')
-  }
+    status: $(this).data("status"),
+  };
   console.log(readiness);
   $.ajax({
-    method: 'PUT',
+    method: "PUT",
     url: `/koalas/${myID}`,
-    data: readiness
-  }).then(function(response){
-    console.log('back from /koalas PUT', response);
-    getKoalas();
-  }).catch( function( err ){
-    console.log(err);
+    data: readiness,
   })
+    .then(function (response) {
+      console.log("back from /koalas PUT", response);
+      getKoalas();
+    })
+    .catch(function (err) {
+      console.log(err);
+    });
 }
 
-function showKoalas(array){
-  console.log( 'in showKoalas' );
-  $('#viewKoalas').empty();
+function showKoalas(array) {
+  console.log("in showKoalas");
+  $("#viewKoalas").empty();
   for (let i = 0; i < array.length; i++) {
     const element = array[i];
-    $('#viewKoalas').append(`
-    <tr>
+    $("#viewKoalas").append(`
+    <tr scope="row">
       <td>${element.name}</td>
       <td>${element.age}</td>
       <td>${element.gender}</td>
@@ -102,24 +98,32 @@ function showKoalas(array){
     </tr>
   `);
   }
-}//end showKoalas
-
+} //end showKoalas
 
 function deleteKoala() {
-  const myID = $(this).data('id');
-  console.log('ID clicked:', myID);
+  swal({
+    text: `are you sure you want to terminate this koala?`,
+    icon: "warning",
+    buttons: ["NO", "YES"],
+    dangerMode: true,
+  }).then((response) => {
+    if (response) {
+      const myID = $(this).data("id");
+      console.log("ID clicked:", myID);
 
-  $.ajax({
-
-    type: 'DELETE',
-    url: `/koalas/${myID}`
-
-  }).then(function(response) {
-      console.log('back from DELETE with:', response);
-      getKoalas();
-  }).catch(function(error) {
-
-      console.log('error in DELETE Route', error);
+      $.ajax({
+        type: "DELETE",
+        url: `/koalas/${myID}`,
+      })
+        .then(function (response) {
+          console.log("back from DELETE with:", response);
+          getKoalas();
+        })
+        .catch(function (error) {
+          console.log("error in DELETE Route", error);
+        });
+    } else {
+      swal("the Koala has been saved!");
+    }
   });
-
-}// end deleteKoala
+} // end deleteKoala
